+++ Makefile.am
@@ -404,8 +404,6 @@ libarchive_test_SOURCES= \
 	libarchive/test/test_archive_read.c \
 	libarchive/test/test_archive_read_add_passphrase.c \
 	libarchive/test/test_archive_read_close_twice.c \
-	libarchive/test/test_archive_read_close_twice_open_fd.c \
-	libarchive/test/test_archive_read_close_twice_open_filename.c \
 	libarchive/test/test_archive_read_multiple_data_objects.c \
 	libarchive/test/test_archive_read_next_header_empty.c \
 	libarchive/test/test_archive_read_next_header_raw.c \
